Rename 'animate' to 'transition'
